Restructure tests to make room for end-to-end tests
