+++ package.json
@@ -10,9 +10,9 @@
     "delete-user": "babel-node scripts/delete-user.js",
     "dev-login": "babel-node scripts/dev-login.js",
     "postinstall": "scripts/deploy-bundle.sh",
-    "test": "npm run linter && npm run test-mocha",
+    "test": "npm run linter && npm run test-functional",
     "linter": "eslint --ext js --ext jsx src test scripts webpack.development.config.babel.js webpack.production.config.babel.js",
-    "test-mocha": "NODE_ENV=test mocha --timeout 4000 --compilers js:babel-core/register",
+    "test-functional": "NODE_ENV=test mocha --timeout 4000 --compilers js:babel-core/register test/functional",
     "seed-database": "babel-node scripts/seed-database.js",
     "create-fake-data": "babel-node scripts/create-fake-data.js",
     "clear-temporary-tables": "babel-node src/kuksa-integration/clear-temporary-tables.js",
