+++ test/functional/registryuser-block-test.js
@@ -1,8 +1,8 @@
-import app from '../src/server/server';
+import app from '../../src/server/server';
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import * as testUtils from './utils/test-utils';
-import { resetDatabase } from '../scripts/seed-database';
+import * as testUtils from '../utils/test-utils';
+import { resetDatabase } from '../../scripts/seed-database';
 import Promise from 'bluebird';
 
 chai.use(chaiAsPromised);
