+++ test/utils/kuksa-integration/fixtures/all/LeirilippukunnatMaksut
