+++ test/utils/kuksa-integration/fixtures/all/Osallistujat
