+++ test/utils/kuksa-integration/fixtures/all/OsallistujatLisatietokentat
