+++ test/utils/kuksa-integration/fixtures/all/OsallistujatLisavalinnat
