+++ test/utils/kuksa-integration/fixtures/all/OsallistujatMaksunTila
