+++ test/utils/kuksa-integration/fixtures/all/RyhmatLisatietokentat
