+++ test/utils/kuksa-integration/fixtures/all/RyhmatLisavalinnat
