+++ test/utils/kuksa-integration/fixtures/all/TapahtumaLisavalinnanPaaryhmat
