+++ test/utils/kuksa-integration/fixtures/all/TapahtumaMaksunPaaryhmat
